Movie findMovie to App component
